requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
csvvalidator==1.2
idna==2.7
isodate==0.6.0
jsonschema==2.5.1
more-itertools==4.3.0
pep8==1.7.0
pluggy==0.8.0
py==1.7.0
pyparsing==2.3.0
pytest==3.10.0
rdflib==4.2.1
requests==2.20.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.11.5
py-multihash==0.2.3